Overhaul of integer types, mostly to use size_t instead of ints. (#22)
We use vector::size_type for the barcode index, based on the BarcodePool. We
generalize the unmatched or ambiguous error code and replace '>= 0' tests of
our now-unsigned integers with the new 'is_barcode_index_ok()' function.

We use size_t's for the string lengths, which is pretty natural for array
indexing. However, we still use regular int's for the number of mismatches
because that should be more than enough. 

For the read counts, we switch to unsigned long longs. This gives us at least
64 bits of counting, which avoids problems when processing very large FASTQ
files involving billions of reads that could overflow 32-bit integers.

We re-use the BarcodeIndex for the trie node pointers. Hopefully this is large
enough, but if it isn't, we detect overflows and throw appropriately. 

For loops over vectors, we use the vector's size_type for most correctness.

Also did some minor edits to the the trie classes so that their Result classes
are more consistent with those in the BarcodeSearch class.