Cache AMBIGUOUS hits from the trie for improved efficiency.
This would still occur with maximum allowed mismatches, so it isn't subject to
the same restrictions on caching as the UNMATCHED case.

We also document that the reported trie/search mismatches should be ignored
when index are UNMATCHED; this allows us to skip sanitizing cached outputs
because we don't have to be exact about the number of reported mismatches.

Fleshed out tests for these new considerations, and updated docstrings.